New API integration: resolutions.

+++ sql/houses/provinces.sql
@@ -1,7 +1,7 @@
 
 SELECT p.province_id,
        p.name AS province_name,
-       COUNT(*) AS total
+       COUNT(*)::int AS total
 FROM houses
 JOIN provinces p USING (province_id)
 GROUP BY p.province_id, p.name;
